Refactor logger for client/server
