+++ next.config.js
@@ -6,7 +6,7 @@ module.exports = {
   eslint: {
     dirs: ['app', 'src']
   },
-  webpack(config, { dev }) {
+  webpack(config, { dev, isServer }) {
     if (!dev) {
       config.devtool = 'source-map'
     }
@@ -14,6 +14,10 @@ module.exports = {
     config.resolve.alias = {
       ...(config.resolve.alias || {}),
       '@': path.resolve(__dirname, 'src'),
+      '@lib/logger': path.resolve(
+        __dirname,
+        isServer ? 'src/lib/logger.server.ts' : 'src/lib/logger.client.ts'
+      ),
     };
     config.module = config.module || { rules: [] }
     config.module.rules.push({
