+++ src/lib/logger.client.ts
@@ -1,0 +1,5 @@
+export const logger = {
+  info: (msg: string) => console.info(msg),
+  debug: (msg: string) => console.debug(msg),
+  error: (msg: string) => console.error(msg),
+}
